#10288CMDB 아이콘 선택 모달에서 불러오는 아이콘이 다르다.(팀장님 피드백 적용)

+++ src/main/kotlin/co/brainz/itsm/image/controller/ImageRestController.kt
@@ -65,6 +65,6 @@ class ImageRestController(private val fileService: AliceFileService) {
         @RequestParam(value = "searchValue", defaultValue = "") searchValue: String,
         @RequestParam(value = "offset", defaultValue = "-1") offset: String
     ): List<AliceImageFileDto> {
-        return fileService.getImageFileList(type, searchValue, offset.toLong())
+        return fileService.getImageFileList(type, searchValue, offset.toInt())
     }
 }
